# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel demo Api Documentation",
        "description": "L5 Swagger OpenApi description of this demo project",
        "contact": {
            "email": "azizul@coderslab.com.bd"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/book/get/all": {
            "get": {
                "tags": [
                    "BOOK"
                ],
                "summary": "Get all books list ",
                "description": "Get all books list from database",
                "operationId": "getBookAll",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found!"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
        "/book/get/all/paginated": {
            "post": {
                "tags": [
                    "BOOK"
                ],
                "summary": "book list paginated",
                "description": "get paginated book list form database",
                "operationId": "getBookPaginated",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "perPage": {
                                        "description": "number of customer per page",
                                        "type": "text"
                                    },
                                    "page": {
                                        "description": "page number",
                                        "type": "text"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "perPage": {
                                        "type": "string",
                                        "example": "15"
                                    },
                                    "page": {
                                        "type": "string",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
        "/book/details/{id}": {
            "get": {
                "tags": [
                    "BOOK"
                ],
                "summary": "Get book Details details",
                "description": "Get book details from database",
                "operationId": "getBookDetails",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of Book to return",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found!"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
        "/book/insert": {
            "post": {
                "tags": [
                    "BOOK"
                ],
                "summary": "insert book into databsae",
                "description": "insert  book into databsae",
                "operationId": "bookInsert",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "name of the book",
                                        "type": "text"
                                    },
                                    "price": {
                                        "description": "price of the book",
                                        "type": "text"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "BOOK1"
                                    },
                                    "price": {
                                        "type": "string",
                                        "example": "10"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
        "/book/update": {
            "post": {
                "tags": [
                    "BOOK"
                ],
                "summary": "update a book into databsae",
                "description": "update book into databsae",
                "operationId": "bookUpdate",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "description": "id of book which will be updated",
                                        "type": "text"
                                    },
                                    "name": {
                                        "description": "name of the book",
                                        "type": "text"
                                    },
                                    "price": {
                                        "description": "price of the book",
                                        "type": "text"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "string",
                                        "example": "1"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "BOOK1"
                                    },
                                    "price": {
                                        "type": "string",
                                        "example": "10"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
        "/book/delete": {
            "post": {
                "tags": [
                    "BOOK"
                ],
                "summary": "delete a book from databsae",
                "description": "delete book from databsae",
                "operationId": "bookDelete",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "description": "id of book which will be updated",
                                        "type": "text"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "string",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Successful Delete operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}